+++ src/main/kotlin/io/github/siyual_park/repository/CategoryRepository.kt
@@ -1,6 +1,8 @@
 package io.github.siyual_park.repository
 
 import io.github.siyual_park.model.category.Category
+import io.github.siyual_park.repository.base.CustomRepository
+import io.github.siyual_park.repository.base.SimpleCustomRepository
 import io.github.siyual_park.repository.specification.CategorySpecification
 import org.springframework.stereotype.Component
 import javax.persistence.EntityManager
