+++ src/main/kotlin/io/github/siyual_park/repository/base/CustomRepository.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.repository
+package io.github.siyual_park.repository.base
 
 interface CustomRepository<T : Any, ID, FACTORY> :
     SpecificationDSLRepository<T, ID, FACTORY>,
