+++ src/main/kotlin/io/github/siyual_park/repository/base/SimpleSpecificationPaginationRepository.kt
@@ -1,17 +1,27 @@
-package io.github.siyual_park.repository
+package io.github.siyual_park.repository.base
 
 import org.springframework.data.repository.NoRepositoryBean
 import javax.persistence.EntityManager
 import kotlin.reflect.KClass
 
 @NoRepositoryBean
-open class SimpleSpecificationRepository<T : Any, ID>(
+open class SimpleSpecificationPaginationRepository<T : Any, ID>(
     crudRepository: CrudRepository<T, ID>,
     clazz: KClass<T>,
     entityManager: EntityManager
-) : SpecificationRepository<T, ID>,
+) : SpecificationPaginationRepository<T, ID>,
     CrudRepository<T, ID> by crudRepository,
     SpecificationRepositoryExpansion<T> by SimpleSpecificationRepositoryExpansion(
+        crudRepository,
+        clazz,
+        entityManager
+    ),
+    PaginationRepositoryExpansion<T> by SimplePaginationRepositoryExpansion(
+        crudRepository,
+        clazz,
+        entityManager
+    ),
+    SpecificationPaginationRepositoryExpansion<T> by SimpleSpecificationPaginationRepositoryExpansion(
         crudRepository,
         clazz,
         entityManager
@@ -19,7 +29,7 @@ open class SimpleSpecificationRepository<T : Any, ID>(
     companion object {
         inline fun <reified T : Any, ID> from(
             entityManager: EntityManager
-        ) = SimpleSpecificationRepository<T, ID>(
+        ) = SimpleSpecificationPaginationRepository<T, ID>(
             SimpleCrudRepository(T::class, entityManager),
             T::class,
             entityManager
