+++ src/main/kotlin/io/github/siyual_park/repository/base/SimpleSpecificationRepository.kt
@@ -1,27 +1,17 @@
-package io.github.siyual_park.repository
+package io.github.siyual_park.repository.base
 
 import org.springframework.data.repository.NoRepositoryBean
 import javax.persistence.EntityManager
 import kotlin.reflect.KClass
 
 @NoRepositoryBean
-open class SimpleSpecificationPaginationRepository<T : Any, ID>(
+open class SimpleSpecificationRepository<T : Any, ID>(
     crudRepository: CrudRepository<T, ID>,
     clazz: KClass<T>,
     entityManager: EntityManager
-) : SpecificationPaginationRepository<T, ID>,
+) : SpecificationRepository<T, ID>,
     CrudRepository<T, ID> by crudRepository,
     SpecificationRepositoryExpansion<T> by SimpleSpecificationRepositoryExpansion(
-        crudRepository,
-        clazz,
-        entityManager
-    ),
-    PaginationRepositoryExpansion<T> by SimplePaginationRepositoryExpansion(
-        crudRepository,
-        clazz,
-        entityManager
-    ),
-    SpecificationPaginationRepositoryExpansion<T> by SimpleSpecificationPaginationRepositoryExpansion(
         crudRepository,
         clazz,
         entityManager
@@ -29,7 +19,7 @@ open class SimpleSpecificationPaginationRepository<T : Any, ID>(
     companion object {
         inline fun <reified T : Any, ID> from(
             entityManager: EntityManager
-        ) = SimpleSpecificationPaginationRepository<T, ID>(
+        ) = SimpleSpecificationRepository<T, ID>(
             SimpleCrudRepository(T::class, entityManager),
             T::class,
             entityManager
